% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applyMask.R
\name{applyMask}
\alias{applyMask}
\title{Applying a Mask (\code{BrcParcellation}) To \code{BrcFmri}}
\usage{
applyMask(fmri, parcellation)
}
\arguments{
\item{fmri}{the \code{BrcFmri} instance}

\item{parcellation}{the \code{BrcParcellation} instance of the same 
dimension as \code{fmri}}
}
\value{
a new \code{BrcFmri} instance
}
\description{
Given a \code{BrcFmri} instance (\code{fmri}) and \code{BrcParcellation}
instance (\code{parcellation}), \code{applyMask} intersects \code{fmri}
by \code{parcellation}. This means all the voxels in \code{parcellation}'s
that were 0 will be zero-ed out in \code{fmri} (i.e., the partition of
that voxel will be set to 0). If no elements of any partition originally 
in \code{fmri} exist anymore after this procedure, its corresponding 
column in \code{fmri} will be removed. On the flip side, if there were
voxel in \code{parcellation} that were non-zero, but its corresponding
voxel in \code{fmri} is empty, the voxel will be assigned an entirely
new parcellation in the output of \code{applyMask}, and its corresponding
column in \code{fmri$data2d} will be the column of all 0's.
}
\details{
Here, \code{fmri} and \code{parcellation} must have the same size.

We call this procedure applying a mask because this procedure does not
respect the partitioning in \code{parcellation}. It only takes into account
which voxels are zero and non-zero in \code{parcellation}.
}

