% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildBrcFmri.R
\name{buildBrcFmri}
\alias{buildBrcFmri}
\title{BrcFmri Builder}
\usage{
buildBrcFmri(data2d, dim3d, id = "", partition = NULL)
}
\arguments{
\item{data2d, }{the 2D fMRI matrix (time x voxels)}

\item{dim3d, }{the dimensions of the 3D fMRI volume}

\item{id, }{an ID string}

\item{partition, }{a factor specifying the 3D fMRI partitions. The default 
(\code{NULL}) will put every 3D fMRI voxel in its own partition. If you 
wish to specify a custom partition, see \code{\link{BrcParcellation}} for 
format details.}
}
\value{
a \code{BrcFmri} instance
}
\description{
\code{buildBrcFmri} is a helper function that quickly build a \code{BrcFmri}
instance. Its main benefit is that it constructs all of the BrcFmri object's
dependencies (a \code{BrcParcellation} and a \code{partition}) for you.
}
\seealso{
\code{\link{BrcFmri}}, \code{\link{BrcParcellation}}
}

