% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildBrcFmri.R
\name{buildBrcParcellation}
\alias{buildBrcParcellation}
\title{\code{BrcParcellation} Builder}
\usage{
buildBrcParcellation(data3d)
}
\arguments{
\item{data3d}{a 3-dimensional matrix}
}
\value{
a \code{BrcParcellation} instance
}
\description{
\code{buildBrcParcellation} is a helper function that quickly build a 
\code{BrcParcellation} instance from a 3-dimensional matrix. Its main 
benefit is that it constructs all of the \code{BrcParcellation} 
object's dependencies (\code{dim3d} and \code{partition}) for you.
}
\details{
Here, the input \code{data3d} is a 3-dimensional matrix, and each
unique value in \code{data3d} will be assigned a unique parcel. The value
0 in \code{data3d} will signal an empty voxel.
}

