% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{convertVoxelIndexto3D}
\alias{convertVoxelIndexto3D}
\title{Convert a Voxel's Position Index to Coordinates (3-dimensional)}
\usage{
convertVoxelIndexto3D(dim3d, idx)
}
\arguments{
\item{dim3d}{a numeric vector of length 3}

\item{idx}{a numeric}
}
\value{
numeric of length 3 representing the 3D coordinates
}
\description{
idx represents the numeric position of a voxel and
dim3d represents the total number of voxels in each direction (x,y,z 
direction) for the fmri. Converts idx into the 3-dimensional
coordinates of the voxel in 3-dimensional array.
}

