% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{data2dTo4d}
\alias{data2dTo4d}
\title{Convert a 2D fMRI data matrix to a 4D fMRI array}
\usage{
data2dTo4d(mat2d, parcellation)
}
\arguments{
\item{mat2d}{the 2D fMRI data matrix (which will usually come from 
\code{BrcFmri$data2d}.)}

\item{parcellation}{a \code{BrcParcellation} instance (which will usually
come from \code{BrcFmri$parcellation}).}
}
\value{
a 4D array containing the series of fMRI volumes.
}
\description{
The 2D data matrix's dimensions are time x voxels. It is suitable for
statistical analysis. The 4D data matrix's dimensions are length x width x
height x time. it is suitable for visualization.
}
\seealso{
\code{\link{data4dTo2d}, the inverse function}
}

