% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{data4dTo2d}
\alias{data4dTo2d}
\title{Convert a 4D array containing a series of fMRI volumes to a 2D fMRI data 
matrix.}
\usage{
data4dTo2d(arr4d)
}
\arguments{
\item{arr4d}{the 4D array containing the series of fMRI volumes}
}
\value{
a 2D fMRI data matrix
}
\description{
The 4D data matrix's dimensions are length x width x height x time. it is
suitable for visualization. The 2D data matrix's dimensions are time x
voxels. It is suitable for statistical analysis.
}

