% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BrcFmri.R
\name{dim4d}
\alias{dim4d}
\title{4D fMRI dimensions}
\usage{
dim4d(mri)
}
\arguments{
\item{mri}{a \code{\link{BrcFmri}} instance}
}
\value{
a 4-element numeric vector containing the \code{BrcFmri}'s
  dimensions.
}
\description{
\code{dim4d} gets the 4D dimensions of a \code{BrcFmri} object.
}
\details{
The first three dimensions are the length, width, and height of each volume. 
The fourth dimension (time) is the scan length.
}

