% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BrcFmri.R
\name{isValid.BrcFmri}
\alias{isValid.BrcFmri}
\title{Checking \code{BrcFmri} instance validity.}
\usage{
\method{isValid}{BrcFmri}(obj)
}
\arguments{
\item{obj}{The \code{BrcFmri} instance to check}
}
\value{
void
}
\description{
\code{isValid} method for class "\code{BrcFmri}".
}
\details{
Fails noisily with a stop message if the \code{BrcFmri}
instance is invalid. Otherwise, nothing happens.
}

