% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BrcParcellation.R
\name{isValid.BrcParcellation}
\alias{isValid.BrcParcellation}
\title{Checking \code{BrcParcellation} instance validity.}
\usage{
\method{isValid}{BrcParcellation}(obj)
}
\arguments{
\item{obj}{The \code{BrcParcellation} instance to check}
}
\value{
void
}
\description{
\code{isValid} method for class "\code{BrcParcellation}.
}
\details{
Fails noisily with a stop message if the \code{BrcParcellation}
instance is invalid. Otherwise, nothing happens.
}

