% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BrcParcellation.R
\name{numParcels}
\alias{numParcels}
\title{Count number of unique parcels in BrcParcellation object}
\usage{
numParcels(parcellation)
}
\arguments{
\item{parcellation}{the BrcParcellation object}
}
\value{
the number of parcels
}
\description{
Count number of unique parcels in BrcParcellation object
}

