% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BrcFmri.R
\name{print.BrcFmri}
\alias{print.BrcFmri}
\title{Printing BrcFmri objects}
\usage{
\method{print}{BrcFmri}(x, ...)
}
\arguments{
\item{x}{a BrcFmri instance}

\item{...}{unused}
}
\description{
\code{print} method for class "\code{BrcFmri}.
}

