% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BrcParcellation.R
\name{print.BrcParcellation}
\alias{print.BrcParcellation}
\title{Printing BrcParcellation objects}
\usage{
\method{print}{BrcParcellation}(x, ...)
}
\arguments{
\item{x}{a BrcParcellation instance}

\item{...}{unused}
}
\description{
\code{print} method for class "\code{BrcParcellation}.
}

