% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{readNifti}
\alias{readNifti}
\title{Reads NIfTI fMRI data}
\usage{
readNifti(file, id = NULL)
}
\arguments{
\item{file}{the file path to read}

\item{id}{the id for the the \code{BrcFmri} output. If left \code{NULL},
the ID will be set as the file path.}
}
\value{
a BrcFmri instance with the data from \code{file}
}
\description{
\code{readNifti} reads a NIfTI formatted fMRI data file into a 
\code{BrcFmri} object.
}

