% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduce.BrcFmri.R
\name{reduce.BrcFmri}
\alias{reduce.BrcFmri}
\title{Reducing function for BrcFmri object given parcellation}
\usage{
\method{reduce}{BrcFmri}(obj, template, func = reduce_mean)
}
\arguments{
\item{obj}{The BrcFmri object to reduce}

\item{template}{The BrcParcellation object to use as a template}

\item{func}{The function to apply to each partition of parcellation in
obj}
}
\value{
A BrcFmri object with the reduced data2d object, the same
id as obj, and the parcellation of template
}
\description{
Reduces the BrcFmri object passed in based on applying a function
over each partition specified by a given BrcParcellation object. 
Both the BrcFmri and BrcParcellation must have the same dimensions.
}

