% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduce.BrcFmri.R
\name{reduce}
\alias{reduce}
\title{Reduces S3's representation}
\usage{
reduce(obj, template, func)
}
\arguments{
\item{obj}{The object to reduce}

\item{template}{The template specifying the regions to reduce obj to}

\item{func}{The function specifying how to do the reduction}
}
\value{
void
}
\description{
Generic function to reduce the S3 object's representation to
a simpler one. Its implementation depends on the specific class
passed in.
}

