% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduce.R
\name{reduce_mean}
\alias{reduce_mean}
\title{Reducing using the mean for reduce.BrcFmri}
\usage{
reduce_mean(mat)
}
\arguments{
\item{mat}{The matrix where each column represents a time series with nrow(mat)
values.}
}
\value{
A vector with nrow(mat) values
}
\description{
If mat is a matrix with no columns, the all 0-vector is returned.
}

