% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduce.R
\name{reduce_pca}
\alias{reduce_pca}
\title{Reducing using the first principal component for reduce.BrcFmri}
\usage{
reduce_pca(mat)
}
\arguments{
\item{mat}{The matrix where each column represents a time series with nrow(mat)
values.}
}
\value{
A vector with nrow(mat) values
}
\description{
If mat is a matrix with no columns, the all 0-vector is returned. Be warned that
the first principal component is not unique, as the negative of the returned
value is also a valid first principal component. We use the first principal
component decided upon by stat::princomp.
}

