% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BrcFmri.R
\name{summary.BrcFmri}
\alias{summary.BrcFmri}
\title{Summarizing BrcFmri objects}
\usage{
\method{summary}{BrcFmri}(object, ...)
}
\arguments{
\item{object}{a BrcFmri instance}

\item{...}{unused}
}
\description{
\code{summary} method for class "\code{BrcFmri}.
}

