% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BrcParcellation.R
\name{summary.BrcParcellation}
\alias{summary.BrcParcellation}
\title{Summarizing BrcParcellation objects}
\usage{
\method{summary}{BrcParcellation}(object, ...)
}
\arguments{
\item{object}{a BrcParcellation instance}

\item{...}{unused}
}
\description{
\code{summary} method for class "\code{BrcParcellation}.
}

