% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{voxel3DToIdx}
\alias{voxel3DToIdx}
\title{Convert a Voxel's Coordinates (3-dimensional) to Position Index}
\usage{
voxel3DToIdx(dim3d, vec)
}
\arguments{
\item{dim3d}{a numeric vector of length 3}

\item{vec}{a numeric vector of length 3}
}
\value{
numeric representing vec's position in an array of dimension dim3d
}
\description{
vec represents the 3-dimensional coordinates of a voxel and
dim3d represents the total number of voxels in each direction (x,y,z 
direction) for the fmri. Converts vec into the numeric that represents
the position of vec in 3-dimensional array.
}

