% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildBrcFmri.R
\name{buildBrcFmri}
\alias{buildBrcFmri}
\title{BrcFmri Builder}
\usage{
buildBrcFmri(data4d, id = "")
}
\arguments{
\item{data4d, }{the 4D fMRI matrix (x-direction x y-direction x z-direction
x time)}

\item{id, }{an ID string}
}
\value{
a \code{BrcFmri} instance
}
\description{
\code{buildBrcFmri} is a helper function that quickly build a \code{BrcFmri}
instance from a 4-dimensional matrix. Its main benefit is that it 
constructs all of the BrcFmri object's dependencies 
(a \code{BrcParcellation} and a \code{partition}) for you. 
Here, data4d is a 4-dimensional matrix where the first 3 dimensions 
refer to the x, y, z coordinates and the 4th dimension refers to time.
}
\details{
Along each 3-dimensional matrix indexed by time, each voxel corresponds to
an x, y, z coordinate. An non-empty voxel is a voxel whose vector along the
4th-dimension (time) is all non-zero. \code{buildBrcFmri} assumes a
singleton parcellation where each non-empty voxel is assigned its own 
parcel.
}
\seealso{
\code{\link{BrcFmri}}, \code{\link{BrcParcellation}}
}

