% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduce.R
\name{reduce}
\alias{reduce}
\title{Reducing function for BrcFmri object given parcellation}
\usage{
reduce(fmri, parcellation, func = reduce_mean)
}
\arguments{
\item{fmri}{The BrcFmri object to reduce}

\item{parcellation}{The BrcParcellation object to use as a parcellation}

\item{func}{The function to apply to each partition of parcellation in
fmri}
}
\value{
A BrcFmri object with the reduced data2d object, the same
id as fmri, and the parcellation of parcellation
}
\description{
Reduces the BrcFmri object passed in based on applying a function
over each partition specified by a given BrcParcellation object. 
Both the BrcFmri and BrcParcellation must have the same dimensions.
}

